testbench: restart kafka if brokers do not come up
kafka tests are plagued by brokers failing to start, which then causes
(false positive) test failures.

We now to detect missing brokers, and do a kafka/zookeepr shutdown and
restart in this case. If that fails again, let's skip the test instead
of failing it - after all, it's not the issue of the to-be-checked code.

see also https://github.com/rsyslog/rsyslog/issues/3057