bugfix imkafka: module makes rsyslog hang if librdkafka hangs
If during shutdown the imkafka worker threads do not terminate
in a timely manner, they make rsyslog terminate slower. If librdkafka
hangs (e.g. when kafka broker is unavailable), this means rsyslog
itself hangs.

The root cause is that the kafka deinitialization was done in
afterRun() entry point, which is expected to be non-blocking. We
have now moved this into the end of the doRun() entry point, which
may block and is cancelled by rsyslog core if it takes to long to
terminate.

closes https://github.com/rsyslog/rsyslog/issues/3154