testbench: improve kafka test stability
zookeeper startup was not actually checked, it was just hoped
that it started within 2 seconds (sleep). On busy systems, this
may be too short, thus results in an improper start and thus can
result in follow-on problems during the test.

We have now corrected it so that we wait until we see a pid. It's
still questionable if this is sufficient to conclude zookeeper is
really ready for processing. But let's take one step after the
other.

see also https://github.com/rsyslog/rsyslog/issues/3057