debug: add instrumentation to analyze shutdown memory leak
valgrind sometimes reports a leak on shutdown and we are still not sure
if this is a rsyslog bug or something in (some versions of) pthreads lib,
which simply should be suppressed. This adds additional debug info, to be
shown in CI runs, to hopefully help clarify the issue.

see also https://github.com/rsyslog/rsyslog/issues/2902