core/wtp: try to fix valgrind memleak report for thread termination
We are trying to solve a memleak report that is probably a false positive.
And there is some indication that this is indeed a false positive:

* https://sourceware.org/bugzilla/show_bug.cgi?id=3087
  (but the posting from 2006 claims cleanup is now done inside pthreads)
* other suppress it as well:
  http://www.mit.edu/afs.new/sipb/project/php/src/mysql-standard-5.0.18-linux-i686/mysql-test/valgrind.supp

Nevertheless, we could not rule yet we have a real leak, caused by a missing
`pthread_join()` or `pthread_detach()`. So we should be sure this is not the
case before suppressing them - especially as we get this error report only
occasionally.

While I not know it is technically correct, this is an interesting posting
(be sure to read all comments as well!):
https://gelorakan.wordpress.com/2007/11/26/pthead_create-valgrind-memory-leak-solved/
... and with this commit we follow the recommendation - let's see ow it works out!

More Info:
We create threads in detached state. We need to do so as threads terminate
after idle period and joining them for the pure joy of releasing resources
is unnecessary complicated.

Checking glibc's pthread code reveals that the leak is indeed related to
thread local storage. But from quick glimpse this seems to be initialization
done in any case. It does not look related to if the user thread is creating
local storage or not.

see also https://github.com/rsyslog/rsyslog/issues/2902