testbench: suppress valgrind error caused by pthreads lib
finally I give up and honestly think this is a problem in pthreads and
not in rsyslog code. See issue below and pervious commit for more
information.

Unfortunately, this will also mask off cases where we do not properly
call pthread_join() albeit it is needed. Neverthless, this bug is
causing so much CI grief that it is definitely worth it.

closes https://github.com/rsyslog/rsyslog/issues/2902