imfile bugfix: newstyle config default does not work on some platforms
when script based configuration is used (which is recommended), "inotify" mode
is always selected as default, except if solaris is detected, in which case
"FEN" is used.

However, there are some systems (like FreeBSD) which neithe support FEN nor
inotify. On these platforms, imfile will alwas fail if not explicitely
configured to "polling".

This patch now checks system capabilities and selects an appropriate
operations mode (the best available).