bugfix omrelp: emit error messages on suspension
The module did not emit error messages. Actually, it would have done
if librelp would emit them (callback is in place), but librelp does
not do. We expect to fix this at the librelp level as well. But to
make this work with all versions of librelp, we have added basic
error messages to omrelp. They may not be totally easy to understand,
but at least provide the librelp error code (that's the best we can
do in this situation).

see also https://github.com/rsyslog/librelp/issues/130
closes https://github.com/rsyslog/rsyslog/issues/3294