testbench: fix omprog-output-capture-mt.sh test failure in Python 3
Make the test program explicitly flush lines written to stdout and
stderr, and do not use stdbuf to force line buffering for stdout.

In Python 2 stderr was unbuffered, but in Python 3 it is block-buffered
when redirected to a file/pipe. Besides, stdbuf seems to have no effect
on Python 3.

This solves one of the problems with this test described in #3361.
The problem ocurred with Python 3 and after the addition of the
wait_file_lines check in v8.40.0. The buffering that Python did caused
the last two lines of the output file to be written at shutdown time.

Other minor fixes and consistency enhancements in other omprog tests.