testbench: add tests for "certless" tcp/tls
This adds a test to ensure that a client without certificate can
connect to a server with certificates. So it is not exactly
"certless".

The prime intent of this test is to match config suggestions given
by log hosting companies (like loggly) and so ensure that we do
not accidently break them. This is espcially important as the
capability for certless clients was not properly documented and
also become forgotten by the rsyslog team.

see also https://github.com/rsyslog/rsyslog/issues/3413