omclickhouse: work-around failed error detection
omclickhouse uses a questionable method to check if a request generated
an error. We have seen the method to fail when we slightly upgraded clickhouse
server in CI testing.

This commit makes the method a bit more reliable without really fixing it.
But it's at least a short-term solution.

This should be changed to a proper status check. I assume such is possible.

see also https://github.com/rsyslog/rsyslog/issues/3485