testbench: fix daqueue-dirty-shutdown test
This test occasionally failed with left-over spool files. As far as we
have analyzed, this is due to the use of an invalid shutdown timeout
(very short) in the second phase of the test. It looks like this is
actually a copy&paste error from phase one. Behaviour of rsyslog was
correct, but the test itself created a false positive.

We have corrected the timeout now and also modernized the test
a bit.

closes https://github.com/rsyslog/rsyslog/issues/2122