imfile: Added new option statefilealwaysremove
Due problems with logrotate and INOTIFY, we added an option
to control how statefiles for moved files are handeled.

By default "statefilealwaysremove" will be off. But if "on"
the statefile for moved files will also be removed.

Otherwise when using logrotate for example, imfile could stop monitoring
rotated files. This problem only happens in INOTIFY mode.

testbench: Added new tests for logrotate and multiple rotates.