omelasticsearch: add support for rebindinterval
Adds support for a new parameter `rebindinterval`
This parameter tells omelasticsearch to close the connection and
reconnect to Elasticsearch after this many operations have been
submitted.  The default value `-1` means that omelasticsearch will
not reconnect.  A value greater than `-1` tells omelasticsearch,
after this many operations have been submitted to Elasticsearch,
to drop the connection and establish a new connection.  This is
useful when rsyslog connects to multiple Elasticsearch nodes
through a router or load balancer, and you need to periodically drop
and reestablish connections to help the router balance the
connections.  Use the counter `rebinds` to monitor the
number of times this has happened.

The es-basic.sh test is made non-racy by the check that all
output is generated before terminating

QUEUE_EMPTY_CHECK_FUNC is called when imdiag *think* the queue is empty. It must
point to a function that verifies all other conditions (here: lines of ouputs and
stats content) are met. If yes, rsyslog is terminated. If no, it is kept running
and the testbench plumbing re-starts the wait cycle by looking at the queue
countes again. This runs until either the condition is met or a timeout occurs.