pmnormalize: fix memory leak, improve tests
This patch fixes a set of problems plus provides more and enhanced
tests for the module.

Most important problem was a memory leak that occured when a message
could not be passed at all. For each message that could not be parsed
memory of at least the size the message is leaked. Depending on
traffic pattern this can quickly lead to OOM. Note, however, that
this leak was never reported - it was discovered as part of code
review.

closes https://github.com/rsyslog/rsyslog/issues/2007