openssl network driver: Added support setting openssl configcommands
We are using the gnutlsPriorityString setting variable, to pass
configuration commands to openssl.
The format is one command per line, command and value separated by equal sign (=).
See this sample:
	gnutlsPriorityString="command=value
	secondcommand=value"

Also added a test that disables certain TLS/SSL Version through this parameter which
causes the connection within the test to fail due wrong tls version number.
This error is checked and if found the test will pass.

Support only on OpenSSL 1.02 or higher.
Testcase will log an error if OpenSSL Versions has not TLSv1.3 support.
This can be ignored. The tests works regardless of this logged error.

closes: https://github.com/rsyslog/rsyslog/issues/3605