testbench bugfix: imfile-symlink failed w/ parallel test run
This test sometimes failed. It used a symlinkl to a hardcoded name
rsyslog-link.*.log. This symlink was created but then disappears.
The reason is that upon (every!) test exit, rsyslog-link.*.log is
deleted. So a parallel test running the exit procedure just at the
"right" time can removed that file.

The bug is that the file name should be created using the tests's
dynamic name. This is done now.

closes https://github.com/rsyslog/rsyslog/issues/3550