core: harden shutdown processing on FreeBSD
root cause seems to be that SIGTERM is delivered differently under
FreeBSD. This causes the main thread to not be awaken, and so it
takes until the next janitor interval to come back to life - which
can be far too long. Fixed this bug explicitley awaking the main
thread.

also

* re-enable test that did not work because of this
* fix invalid message on SIGTERM in debug log