tcpsrv: Wait for workers to finish before terminating
The TCP server may still have workers running when its main thread
terminates, and the workers may be using state that is liable to disappear
during the shutdown process.

This fixes a segfault in hashtable_insert() which will occur if a worker
thread processes a connection while rsyslog is shutting down, and attempts
to use the DNS cache after it has been deinitialized (and the hashtable
destroyed) by dnscacheDeinit().

Signed-off-by: John Brooks <jbrooks@ciena.com>