save work state --> take new approach
thought error: if action is suspended, we CAN NOT retry, so we do not
try to do this. The best thing we can do in this case is to write the
error file, but that's it. So for the suspend-via-file we should make
sure we just check at begin of batch.