queue: fixes issue when DiscardMark reached
When DiscardMark is configured for a queue and DiskcardMark elements inserted it is possible for lower priority items to not be discarded. This causes lower priority items to become "stuck" in the queue. This updates the element state to DISCARDED if the queue is full and will delete them because they _have_ been processed.