omelasticsearch bugfix: segfault on unknown retryRuleset
omelasticsearch does some "interesting tricks" for an output module.
This causes a segfault if the retryRuleset is now known.

The action module interface currently expects that all config errors
be detected during instance creation. Instead omelasticsearch defers
the retry ruleset check to a later state. The reason is that it wants
to support the use the same rulesetname it is defined in - and this
is not yet available at action parsing.

We fix this by ensuring that any deleted instance is properly unlinked
from the instance list. One may argue the module interface should get
upgrade for such cases, but this is a longer-term approach.

replaces https://github.com/rsyslog/rsyslog/pull/3796
closes https://github.com/rsyslog/rsyslog/pull/3796