Do not close /dev/urandom file descriptor after fork on Linux systems
This patch updates prepareBackground() in tools/rsyslogd.c to check that an open
file descriptor is not being used for random number generation before closing it
as an unnecessary open file. This fixes an issue on Linux systems where the file
descriptor obtained for /dev/urandom by seedRandomNumber() in runtime/srutils.c
was being closed after the fork. This could be observed in procfs, where
/proc/<PID>/fd/ would show no open descriptors to /dev/urandom in the forked
process. With this fix applied, procfs confirms that the forked rsyslog process
does retain a valid file descriptor to /dev/urandom.

I found that this issue led to rsyslog intermittently hanging during seedIV()
in runtime/libgcry.c. After the fork, the closed file descriptor number tended
to get re-assigned. randomNumber() would then read from an incorrect (although
still valid) file descriptor, and could block (depending on the state of that
file descriptor). This gave rise to the intermittent hang that I observed.

Signed-off-by: Simon Haggett <simon.haggett@gmail.com>