build system: change --enable-imfile-tests default to "yes"
This was accidentally set to "no". Test for imfile should by
default run when imfile is enabled.

We also temporarily disable the failing tests - this makes
it posible to run the remaining tests during CI and prevent
future errors. That is better than nothing. Tests are to be
re-enabled when the imfile bug mentioned below is fixed.

see also https://github.com/rsyslog/rsyslog/issues/4120

Signed-off-by: Rainer Gerhards <rgerhards@adiscon.com>