devel tools: show stalled tests from "make check"
CI systems cancel out runs that are stalled. Unfortunately, it is very
hard to find the test that caused the stall. This commit extends the
toolset to detect logs without a matching .trs file, which is a strong
indication of an aborted test.

This alos modernizes some tests in order to remove sample *.log files
which would otherwise always clutter log output.