act_obj_add should not return RS_RET_OK if open call fails
If we configure more number of files than the fd limit for log rotation,
then in function act_obj_add, it tries to open each of the file
and because the fd is exhasuted, the open system call fails.

So, in act_obj_add even if open fails we are returning RS_RET_OK
to process_symlink, which then starts accessing chld->active->name and
chld->active->basename. Here it crashes as chld->Active is NULL. Its supposed to
be initialized in act_obj_add which bailed out as open call failed but still returning
RS_RET_OK

So, this situation can be handled if act_obj_add doesnt return RS_RET_OK if
open fails.

Added test for the same behaviour.

So, in test we keep creating more files for imfile to parse. Once we see the FD is
exhasuted, we stop creating files.

I see once the FD exhasuted, we will hit this condition in tcpsrv.c
https://github.com/rsyslog/rsyslog/blob/3574f39a4b18e84c3bc5a92f1ca8760ff7985a40/runtime/tcpsrv.c#L987

So, the process stays in the while loop until timeout. Hence, cleaning up the files in
the test so that rsyslog could process the further workset.

Fixes #4173

Signed-off-by: Tapas Kundu <tkundu@vmware.com>