testbench: add some diagnostic info to analyze false positives
We have some false positives with imfile checks and this commit
both improves the testbench framework slightly and adds debug
info.

The debug info should be remove when we finally find the cause
of the issue, but it does not hurt if it stays for a quite a
while. Thus we can analyze the false positives over an extended
period of time - what is what it looks like we need to do to
find the root cause.